chore: move nargo-specific functions from `fm` to `nargo`
