+++ crates/nargo/Cargo.toml
@@ -20,7 +20,6 @@ iter-extended.workspace = true
 noirc_driver.workspace = true
 noirc_frontend.workspace = true
 noirc_abi.workspace = true
-fm.workspace = true
 acvm.workspace = true
 cfg-if.workspace = true
 toml.workspace = true
