+++ crates/nargo/src/workspace.rs
@@ -1,4 +1,4 @@
 // We will say that a cargo unit must contain either a binary or a library
-// Then we use workspace to allow more than one. In the future, do not allow there to be 
+// Then we use workspace to allow more than one. In the future, do not allow there to be
 // both a binary and a library.
 // - library will be default
